Invalid DateTime

+++ -Wk2_Prac2/spr/Move.h
@@ -5,7 +5,7 @@
 class Move {
  protected:
   std::string NameOfMove = " ";
-  int id;
+  float id;
 
  public:
   Move(/* args */);
